reorg source to isolate all tf interfaceing to to tf_loader and tf_utils
