+++ tf2onnx/__init__.py
@@ -6,8 +6,10 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-__all__ = ["utils", "graph_matcher", "graph", "graph_builder", "loader", "tfonnx", "shape_inference", "schemas"]
+__all__ = ["utils", "graph_matcher", "graph", "graph_builder",
+           "tfonnx", "shape_inference", "schemas", "tf_utils", "tf_loader"]
 
 from .version import version as __version__
 from . import verbose_logging as logging
 from tf2onnx import tfonnx, utils, graph, graph_builder, graph_matcher, shape_inference, schemas  # pylint: disable=wrong-import-order
+#from tf2onnx import tf_utils, tf_loader
