+++ tf2onnx/shape_inference.py
@@ -14,6 +14,7 @@ from collections import defaultdict
 
 import tensorflow as tf
 from tf2onnx import utils
+from tf2onnx.tf_utils import get_tf_tensor_shape, get_tf_const_value, get_tf_shape_attr, get_tf_version
 
 # pylint: disable=logging-not-lazy,missing-docstring,consider-swap-variables
 
@@ -53,7 +54,7 @@ def check_shape_for_tf_graph(tf_graph):
     op_outputs_mapping_none_shape = defaultdict(list)
     for op in tf_graph.get_operations():
         for out in op.outputs:
-            if utils.get_tf_tensor_shape(out) is None:
+            if get_tf_tensor_shape(out) is None:
                 op_outputs_mapping_none_shape[op.name].append(out.name)
     return op_outputs_mapping_none_shape
 
@@ -61,7 +62,7 @@ def check_shape_for_tf_graph(tf_graph):
 def reload_tf_graph(tf_graph):
     """Invoke tensorflow cpp shape inference by reloading graph_def."""
     # invoke c api if tf version is below 1.8
-    if utils.get_tf_version() < LooseVersion("1.8"):
+    if get_tf_version() < LooseVersion("1.8"):
         logger.debug(
             "On TF < 1.8, graph is constructed by python API, " \
             "which doesn't invoke shape inference, please set " \
@@ -98,14 +99,14 @@ def infer_shape_for_graph(tf_graph):
 
 
 def infer_shape_for_op(op):
-    has_unknown_output_shape = any(utils.get_tf_tensor_shape(out) is None for out in op.outputs)
+    has_unknown_output_shape = any(get_tf_tensor_shape(out) is None for out in op.outputs)
 
     if not has_unknown_output_shape:
         return False
 
     if op.type == "Placeholder":
         # if placeholder shape is not found, try to get it from "shape" attribute.
-        attr_shape = utils.get_tf_shape_attr(op)
+        attr_shape = get_tf_shape_attr(op)
         if attr_shape is not None:
             new_shape = list(attr_shape)
             op.outputs[0].set_shape(new_shape)
@@ -116,8 +117,8 @@ def infer_shape_for_op(op):
         return True
 
     if op.type == "Merge":
-        s1 = utils.get_tf_tensor_shape(op.inputs[0])
-        s2 = utils.get_tf_tensor_shape(op.inputs[1])
+        s1 = get_tf_tensor_shape(op.inputs[0])
+        s2 = get_tf_tensor_shape(op.inputs[1])
         new_shape = None
         if s1 is None and s2 is None:
             return False
@@ -151,7 +152,7 @@ def infer_shape_for_op(op):
         return True
 
     if op.type == "Switch":
-        new_shape = utils.get_tf_tensor_shape(op.inputs[0])
+        new_shape = get_tf_tensor_shape(op.inputs[0])
         if new_shape is not None:
             op.outputs[0].set_shape(new_shape)
             op.outputs[1].set_shape(new_shape)
@@ -161,7 +162,7 @@ def infer_shape_for_op(op):
         return False
 
     if op.type == "Enter":
-        new_shape = utils.get_tf_tensor_shape(op.inputs[0])
+        new_shape = get_tf_tensor_shape(op.inputs[0])
         if new_shape is not None:
             op.outputs[0].set_shape(new_shape)
             logger.debug("set [%s] with new shape %s", op.outputs[0].name, new_shape)
@@ -184,7 +185,7 @@ def infer_shape_for_op(op):
         if not tensor_array_write_op:
             return False
         # get TensorArray shape from input tensor of the found TensorArrayWrite op
-        shape = utils.get_tf_tensor_shape(tensor_array_write_op.inputs[2])
+        shape = get_tf_tensor_shape(tensor_array_write_op.inputs[2])
         # update TensorArray's shape info
         if shape is not None:
             new_shape = [None] + shape
@@ -206,7 +207,7 @@ def infer_shape_for_op(op):
         if scatter_op.type != "TensorArrayScatterV3":
             return False
 
-        value_shape_before_scatter = utils.get_tf_tensor_shape(scatter_op.inputs[2])
+        value_shape_before_scatter = get_tf_tensor_shape(scatter_op.inputs[2])
         if value_shape_before_scatter is None:
             return False
 
@@ -306,8 +307,8 @@ def infer_shape_for_op_legacy(op):
     # invoke tf shape inference first
     infer_shape_for_op(op)
 
-    has_unknown_input_shape = any(utils.get_tf_tensor_shape(inp) is None for inp in op.inputs)
-    has_unknown_output_shape = any(utils.get_tf_tensor_shape(out) is None for out in op.outputs)
+    has_unknown_input_shape = any(get_tf_tensor_shape(inp) is None for inp in op.inputs)
+    has_unknown_output_shape = any(get_tf_tensor_shape(out) is None for out in op.outputs)
 
     # an input shape may be inferred from op output or other input shapes
     # try to infer it first
@@ -327,7 +328,7 @@ def infer_shape_for_op_legacy(op):
     are_all_input_shape_ready = True
     no_shape = []
     for i in op.inputs:
-        if utils.get_tf_tensor_shape(i) is None:
+        if get_tf_tensor_shape(i) is None:
             are_all_input_shape_ready = False
             no_shape.append(i.name)
 
@@ -350,15 +351,15 @@ def infer_shape_for_op_legacy(op):
     if op.type == "Gather":
         # uses the follwing link to know how to infer shape of output
         # https://www.tensorflow.org/api_docs/python/tf/gather
-        shape_params = utils.get_tf_tensor_shape(op.inputs[0])
-        shape_indices = utils.get_tf_tensor_shape(op.inputs[1])
+        shape_params = get_tf_tensor_shape(op.inputs[0])
+        shape_indices = get_tf_tensor_shape(op.inputs[1])
         # gather can only have 2 inputs
         # https://www.tensorflow.org/api_docs/cc/class/tensorflow/ops/gather.html
         if len(op.inputs) == 3:
             axis_op = op.inputs[2].op
             if not utils.is_tf_const_op(axis_op):
                 return False
-            axis = utils.get_tf_const_value(axis_op)
+            axis = get_tf_const_value(axis_op)
         else:
             axis = 0
 
@@ -370,11 +371,11 @@ def infer_shape_for_op_legacy(op):
         axis_op = op.inputs[1].op
         if not utils.is_tf_const_op(axis_op):
             return False
-        axis = utils.get_tf_const_value(axis_op)
+        axis = get_tf_const_value(axis_op)
         if not isinstance(axis, list):
             axis = [axis]
         keep_dims = op.get_attr("keep_dims")
-        shape = utils.get_tf_tensor_shape(op.inputs[0])
+        shape = get_tf_tensor_shape(op.inputs[0])
         for i, _ in enumerate(axis):
             if axis[i] < 0:
                 axis[i] += len(shape)
@@ -393,12 +394,12 @@ def infer_shape_for_op_legacy(op):
 
     if op.type == "ExpandDims":
         # https://www.tensorflow.org/api_docs/python/tf/expand_dims
-        input_shape = utils.get_tf_tensor_shape(op.inputs[0])
+        input_shape = get_tf_tensor_shape(op.inputs[0])
         dim_op = op.inputs[1].op
         if input_shape is None or not utils.is_tf_const_op(dim_op):
             return False
 
-        dim = utils.get_tf_const_value(dim_op)
+        dim = get_tf_const_value(dim_op)
         if dim < 0:
             dim = dim + len(input_shape) + 1
 
@@ -408,7 +409,7 @@ def infer_shape_for_op_legacy(op):
         return True
 
     if op.type == "Unpack":
-        input_shape = utils.get_tf_tensor_shape(op.inputs[0])
+        input_shape = get_tf_tensor_shape(op.inputs[0])
         if input_shape is None:
             return False
 
@@ -425,7 +426,7 @@ def infer_shape_for_op_legacy(op):
 
     if op.type in ["Minimum", "Maximum"]:
         # ops that are elementwise and support broadcasting
-        input_shapes = [utils.get_tf_tensor_shape(op) for op in op.inputs]
+        input_shapes = [get_tf_tensor_shape(op) for op in op.inputs]
         new_shape = broadcast_shape_inference(*input_shapes)
         op.outputs[0].set_shape(new_shape)
         return True
@@ -435,8 +436,8 @@ def infer_shape_for_op_legacy(op):
 
 def infer_input_shapes(op):
     if op.type == "Select":
-        shape_t = utils.get_tf_tensor_shape(op.inputs[1])
-        shape_e = utils.get_tf_tensor_shape(op.inputs[2])
+        shape_t = get_tf_tensor_shape(op.inputs[1])
+        shape_e = get_tf_tensor_shape(op.inputs[2])
         # copy shape if t OR e does not have a shape, no update if t AND e both have shapes
         if shape_t is None or shape_e is None:
             new_shape = shape_t or shape_e
@@ -453,7 +454,7 @@ def infer_output_shapes_with_partial_inputs(op):
     # so only partial(at least one) input shapes need to be known to infer output shape of concat op
     if utils.is_tf_concat_op(op):
         data_inputs = op.inputs[:-1]
-        input_shapes = [utils.get_tf_tensor_shape(inp) for inp in data_inputs]
+        input_shapes = [get_tf_tensor_shape(inp) for inp in data_inputs]
         input_shapes = [shape for shape in input_shapes if shape is not None]
         if not input_shapes:
             logger.debug("all input shapes of concat op %s are None, can't infer its output shape", op.name)
@@ -466,7 +467,7 @@ def infer_output_shapes_with_partial_inputs(op):
             op.outputs[0].set_shape([-1] * rank)
             return True
 
-        axis = utils.get_tf_const_value(axis_op)
+        axis = get_tf_const_value(axis_op)
         axis = axis if axis >= 0 else axis + rank
         new_shape[axis] = -1
         if len(input_shapes) == len(data_inputs):  # all input shapes are known
@@ -480,9 +481,9 @@ def infer_output_shapes_with_partial_inputs(op):
         return True
 
     if op.type == "Select":
-        new_shape = utils.get_tf_tensor_shape(op.inputs[1])
+        new_shape = get_tf_tensor_shape(op.inputs[1])
         if new_shape is None:
-            new_shape = utils.get_tf_tensor_shape(op.inputs[2])
+            new_shape = get_tf_tensor_shape(op.inputs[2])
         if new_shape is not None:
             op.outputs[0].set_shape(new_shape)
             op.inputs[1].set_shape(new_shape)
@@ -495,7 +496,7 @@ def infer_output_shapes_with_partial_inputs(op):
         axis = op.get_attr("axis")
         input_shape = None
         for i in op.inputs:
-            s = utils.get_tf_tensor_shape(i)
+            s = get_tf_tensor_shape(i)
             if s is not None:
                 input_shape = s
                 break
@@ -504,7 +505,7 @@ def infer_output_shapes_with_partial_inputs(op):
         if axis < 0:
             axis += len(input_shape)
         for i in op.inputs:
-            if not utils.get_tf_tensor_shape(i):
+            if not get_tf_tensor_shape(i):
                 i.set_shape(input_shape)
                 logger.debug("set [%s] with new shape %s", i.name, input_shape)
         new_shape = input_shape[:axis] + [len(op.inputs)] + input_shape[axis:]
@@ -514,9 +515,9 @@ def infer_output_shapes_with_partial_inputs(op):
 
     if op.type == "Pow":
         # https://www.tensorflow.org/api_docs/cc/class/tensorflow/ops/pow
-        new_shape = utils.get_tf_tensor_shape(op.inputs[0])
+        new_shape = get_tf_tensor_shape(op.inputs[0])
         if new_shape is None:
-            new_shape = utils.get_tf_tensor_shape(op.inputs[1])
+            new_shape = get_tf_tensor_shape(op.inputs[1])
         if new_shape is not None:
             op.outputs[0].set_shape(new_shape)
             logger.debug("set [%s] with new shape %s", op.outputs[0].name, new_shape)
@@ -527,7 +528,7 @@ def infer_output_shapes_with_partial_inputs(op):
 
 
 def set_shape_from_input(input_tensor, output_tensor):
-    new_shape = utils.get_tf_tensor_shape(input_tensor)
+    new_shape = get_tf_tensor_shape(input_tensor)
     if new_shape is not None:
         output_tensor.set_shape(new_shape)
         logger.debug("set [%s] with new shape %s", output_tensor.name, new_shape)
@@ -536,8 +537,8 @@ def set_shape_from_input(input_tensor, output_tensor):
 
 
 def set_shape_from_inputs_broadcast(input_tensors, output_tensor):
-    s1 = utils.get_tf_tensor_shape(input_tensors[0])
-    s2 = utils.get_tf_tensor_shape(input_tensors[1])
+    s1 = get_tf_tensor_shape(input_tensors[0])
+    s2 = get_tf_tensor_shape(input_tensors[1])
     new_shape = broadcast_shape_inference(s1, s2)
     if new_shape is not None:
         output_tensor.set_shape(new_shape)
